backend/requirements.txt
chromadb>=0.4.0
transformers>=4.30.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
google-adk>=0.1.0
litellm>=1.65.5
langchain-community>=0.3.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi
uvicorn
openai-whisper
torchaudio